requirements.txt
streamlit
pymupdf
groq
google-genai
python-dotenv
nbformat